Lift the default config into a factory in @meru/shared
The config every install starts from lived as an object literal inside the
electron-store constructor, so nothing outside the main process could read
it. Three values held it there: a generated account id, the downloads
directory, and the macOS check behind the tray icon.

A factory taking those three leaves the main process with one source of
truth and lets anything rendering the app's components outside Electron
read the defaults the app ships with rather than a copy of them.
DEFAULT_WINDOW_STATE_BOUNDS moves along with it, since it depends on
nothing.
