+++ packages/app/config.ts
@@ -1,5 +1,6 @@
 import { randomUUID } from "node:crypto";
 import { is, platform } from "@electron-toolkit/utils";
+import { createDefaultConfig } from "@meru/shared/config";
 import type { SavedTab } from "@meru/shared/schemas";
 import type { Config } from "@meru/shared/types";
 import { app } from "electron";
@@ -10,128 +11,14 @@ type LegacySavedTab = SavedTab & {
   persistence?: "pinned" | "bookmarked";
 };
 
-export const DEFAULT_WINDOW_STATE_BOUNDS = {
-  width: 1280,
-  height: 800,
-  x: undefined,
-  y: undefined,
-};
-
 export const config = new Store<Config>({
   name: is.dev ? "config.dev" : "config",
   accessPropertiesByDotNotation: false,
-  defaults: {
-    accounts: [
-      {
-        id: randomUUID(),
-        label: "Default",
-        color: null,
-        selected: true,
-        notifications: true,
-        gmail: {
-          unreadBadge: true,
-          delegatedAccountId: null,
-          unifiedInbox: true,
-        },
-        workspaceApps: {
-          savedTabs: [],
-          bookmarks: [],
-        },
-      },
-    ],
-    "accounts.unreadBadge": true,
-    "app.hardwareAcceleration": true,
-    launchMinimized: false,
-    launchAtLogin: false,
-    resetApp: false,
-    theme: "system",
-    licenseKey: null,
-    customUserAgent: false,
-    "dock.enabled": true,
-    "dock.unreadBadge": true,
-    "externalLinks.confirm": true,
-    "externalLinks.trustedHosts": [],
-    "downloads.saveAs": false,
-    "downloads.openFolderWhenDone": false,
-    "downloads.location": app.getPath("downloads"),
-    "downloads.history": [],
-    "notifications.enabled": true,
-    "notifications.showSender": true,
-    "notifications.showSubject": true,
-    "notifications.showSummary": true,
-    "notifications.playSound": true,
-    "notifications.allowFromWorkspaceApps": true,
-    "notifications.sound": "linen",
-    "notifications.volume": 1,
-    "notifications.downloadCompleted": true,
-    "notifications.onClickDownloadCompleted": "showInFolder",
-    "notifications.times": [],
-    "updates.autoCheck": true,
-    "updates.showNotifications": true,
-    "updates.channel": "stable",
-    "blocker.enabled": true,
-    "blocker.ads": true,
-    "blocker.tracking": true,
-    "tray.enabled": !platform.isMacOS,
-    "tray.iconColor": "system",
-    "tray.unreadCount": true,
-    "tray.selectAccountWithUnread": false,
-    "gmail.hideGmailLogo": true,
-    "gmail.hideInboxFooter": true,
-    "gmail.hideOutOfOfficeBanner": false,
-    "gmail.hidePromoBanner": true,
-    "gmail.hideUpgradeButton": true,
-    "gmail.reverseConversation": false,
-    "gmail.savedSearches": [],
-    "gmail.labelColors": [],
-    "gmail.unreadCountPreference": "inbox",
-    "gmail.openComposeInNewWindow": false,
-    "gmail.showSenderIcons": true,
-    "gmail.moveAttachmentsToTop": false,
-    "gmail.closeComposeWindowAfterSend": false,
-    "gmail.replyForwardInPopOut": false,
-    "gmail.extendDarkTheme": false,
-    "gmail.inboxCategoriesToMonitor": "primary",
-    "screenShare.useSystemPicker": true,
-    "window.lastState": {
-      bounds: DEFAULT_WINDOW_STATE_BOUNDS,
-      fullscreen: false,
-      maximized: false,
-    },
-    "window.restrictMinimumSize": true,
-    "trial.expired": false,
-    "workspaceApps.openInApp": true,
-    "workspaceApps.openInAppExcludedApps": [],
-    "workspaceApps.mode": "tabs",
-    "workspaceApps.launcherApps": [],
-    "workspaceApps.launcherDisplay": "auto",
-    "workspaceApps.launcherAndBookmarksPlacement": "auto",
-    "workspaceApps.showAccountColor": true,
-    "workspaceApps.showAccountLabel": true,
-    "workspaceApps.persistZoom": true,
-    "workspaceApps.zoomFactors": {},
-    "workspaceApps.hidePasskeyDialog": false,
-    "workspaceApps.hibernation": "selected",
-    "workspaceApps.hibernationTimeout": "1h",
-    "verificationCodes.autoCopy": false,
-    "verificationCodes.copyMode": "immediately",
-    "verificationCodes.autoDelete": false,
-    "verificationCodes.autoMarkAsRead": false,
-    "doNotDisturb.enabled": false,
-    "doNotDisturb.duration": null,
-    "doNotDisturb.until": null,
-    "unifiedInbox.enabled": true,
-    "unifiedInbox.showSenderIcons": true,
-    "unifiedInbox.rowsPerPage": 10,
-    "spellchecker.languages": [],
-    "verticalTabs.showWindows": true,
-    "verticalTabs.width": "auto",
-    "verticalTabs.showWidthToggle": true,
-    "verticalTabs.hideUnreadBadgeWhenActive": false,
-    "verticalTabs.showAppLinksBadge": true,
-    "extensions.installed": [],
-    "extensions.showTitlebarButton": false,
-  },
+  defaults: createDefaultConfig({
+    accountId: randomUUID(),
+    downloadsLocation: app.getPath("downloads"),
+    trayEnabled: !platform.isMacOS,
+  }),
   migrations: {
     ">=3.4.0": (store) => {
       // @ts-expect-error: `showDockIcon` is now 'dock.enabled'
